backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
email-validator
psycopg2-binary
bcrypt==4.0.1
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
alembic
dotenv
pytest
pytest-asyncio
httpx
aiohttp
redis[asyncio]